Make parser a phase
Take care that parser errors do not stop computation, and that
printing after parser prints untyped trees.
